enh: rename checklist endpoints + fix get block pod endpoint
