+++ lib/golang/endpoints/keys.3.checklists/keys.3.4.delete_key_checklist.go
