+++ lib/golang/server.constants.go
@@ -56,15 +56,15 @@ const (
 )
 
 const (
-	RouteBlocksGetBlockChecklists     = "blocks/%s/checklists?keyId=%s"
-	RouteBlocksAddBlockChecklist      = "blocks/%s/checklists?keyId=%s"
-	RouteBlocksReorderBlockChecklists = "blocks/%s/checklists/reorder?keyId=%s"
-	RouteBlocksRenameChecklistTitle   = "blocks/%s/checklists/%s?keyId=%s"
-	RouteBlocksDeleteChecklist        = "blocks/%s/checklists/%s?keyId=%s"
-	RouteBlocksAddChecklistItem       = "blocks/%s/checklists/%s/checklist-items?keyId=%s"
-	RouteBlocksUpdateChecklistItem    = "blocks/%s/checklists/%s/checklist-items/%s?keyId=%s"
-	RouteBlocksDeleteChecklistItem    = "blocks/%s/checklists/%s/checklist-items/%s?keyId=%s"
-	RouteBlocksReorderChecklistItems  = "blocks/%s/checklists/%s/checklist-items/reorder?keyId=%s"
+	RouteBlocksGetBlockChecklists         = "blocks/%s/checklists?keyId=%s"
+	RouteBlocksAddBlockChecklist          = "blocks/%s/checklists?keyId=%s"
+	RouteBlocksReorderBlockChecklists     = "blocks/%s/checklists/reorder?keyId=%s"
+	RouteBlocksDeleteBlockChecklist       = "blocks/%s/checklists/%s?keyId=%s"
+	RouteBlocksRenameBlockChecklist       = "blocks/%s/checklists/%s?keyId=%s"
+	RouteBlocksAddBlockChecklistItem      = "blocks/%s/checklists/%s/checklist-items?keyId=%s"
+	RouteBlocksUpdateBlockChecklistItem   = "blocks/%s/checklists/%s/checklist-items/%s?keyId=%s"
+	RouteBlocksDeleteBlockChecklistItem   = "blocks/%s/checklists/%s/checklist-items/%s?keyId=%s"
+	RouteBlocksReorderBlockChecklistItems = "blocks/%s/checklists/%s/checklist-items/reorder?keyId=%s"
 )
 
 const (
@@ -285,8 +285,8 @@ const (
 	RouteKeysGetKeyChecklists         = "keys/%s/checklists"
 	RouteKeysAddKeyChecklist          = "keys/%s/checklists"
 	RouteKeysReorderKeyChecklists     = "keys/%s/checklists/reorder"
-	RouteKeysRenameChecklist          = "keys/%s/checklists/%s"
 	RouteKeysDeleteKeyChecklist       = "keys/%s/checklists/%s"
+	RouteKeysRenameKeyChecklist       = "keys/%s/checklists/%s"
 	RouteKeysAddKeyChecklistItem      = "keys/%s/checklists/%s/checklist-items"
 	RouteKeysUpdateKeyChecklistItem   = "keys/%s/checklists/%s/checklist-items/%s"
 	RouteKeysDeleteKeyChecklistItem   = "keys/%s/checklists/%s/checklist-items/%s"
